Rename to newResidentAndResidencySchema
