+++ client/views/residencies/form/form.html
@@ -46,7 +46,7 @@
 
             {{/ autoForm }}
           {{ else }}
-            {{> quickForm schema=newResidentFormSchema id="new-resident-and-residency" type="method" }}
+            {{> quickForm schema=newResidentAndResidencySchema id="new-resident-and-residency" type="method" }}
           {{/ if }}
         </div> <!-- / .modal-body -->
         {{ else }}
